pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "shieldlm"
version = "0.1.0"
description = "Fast, deployment-aware prompt injection detection classifier"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Dimiter Milushev"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "datasets>=2.16.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "pyarrow>=14.0.0",
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
train = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "accelerate>=0.27.0",
    "peft>=0.10.0",
    "bitsandbytes>=0.43.0",
    "evaluate>=0.4.0",
    "scipy>=1.11.0",
]
eval = [
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tabulate>=0.9.0",
]
serve = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
